step 26

+++ src/main/java/com/ll/sb_re231120/domain/article/article/controller/ArticleController.java
@@ -4,10 +4,12 @@ import com.ll.sb_re231120.domain.article.article.Service.ArticleService;
 import com.ll.sb_re231120.domain.article.article.entity.Article;
 import com.ll.sb_re231120.domain.member.member.service.MemberService;
 import com.ll.sb_re231120.global.rq.Rq;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Data;
 import lombok.RequiredArgsConstructor;
+import lombok.SneakyThrows;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.validation.annotation.Validated;
@@ -42,6 +44,8 @@ public class ArticleController {
     }
     @GetMapping("/article/write")
     String showWrite() {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         return "article/article/write";
     }
     @Data
@@ -52,12 +56,17 @@ public class ArticleController {
         private String body;
     }
     @PostMapping("/article/write")
-    String write(@Valid WriteForm writeForm) {
+    @SneakyThrows
+    String write(@Valid WriteForm writeForm, HttpServletRequest req) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         Article article = articleService.write(writeForm.title, writeForm.body);
         return rq.redirect("/article/list", "%d번 게시물 생성되었습니다.".formatted(article.getId()));
     }
     @GetMapping("/article/modify/{id}")
     String showModify(Model model, @PathVariable long id) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         Article article = articleService.findById(id).get();
 
         model.addAttribute("article", article);
@@ -75,6 +84,8 @@ public class ArticleController {
 
     @PostMapping("/article/modify/{id}")
     String modify(@PathVariable long id, @Valid ModifyForm modifyForm) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         articleService.modify(id, modifyForm.title, modifyForm.body);
 
         return rq.redirect("/article/list", "%d번 게시물 수정되었습니다.".formatted(id));
@@ -82,6 +93,8 @@ public class ArticleController {
 
     @GetMapping("/article/delete/{id}")
     String delete(@PathVariable long id) {
+        if (!rq.isLogined()) throw new RuntimeException("로그인 후 이용해주세요.");
+
         articleService.delete(id);
 
         return rq.redirect("/article/list", "%d번 게시물 삭제되었습니다.".formatted(id));
